src/test/javascript/spec/app/entities/concept/concept-update.component.spec.ts
import { ComponentFixture, TestBed, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { FormBuilder } from '@angular/forms';
import { of } from 'rxjs';

import { DevpctTestModule } from '../../../test.module';
import { ConceptUpdateComponent } from 'app/entities/concept/concept-update.component';
import { ConceptService } from 'app/entities/concept/concept.service';
import { Concept } from 'app/shared/model/concept.model';

describe('Component Tests', () => {
  describe('Concept Management Update Component', () => {
    let comp: ConceptUpdateComponent;
    let fixture: ComponentFixture<ConceptUpdateComponent>;
    let service: ConceptService;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [DevpctTestModule],
        declarations: [ConceptUpdateComponent],
        providers: [FormBuilder],
      })
        .overrideTemplate(ConceptUpdateComponent, '')
        .compileComponents();

      fixture = TestBed.createComponent(ConceptUpdateComponent);
      comp = fixture.componentInstance;
      service = fixture.debugElement.injector.get(ConceptService);
    });

    describe('save', () => {
      it('Should call update service on save for existing entity', fakeAsync(() => {
        // GIVEN
        const entity = new Concept(123);
        spyOn(service, 'update').and.returnValue(of(new HttpResponse({ body: entity })));
        comp.updateForm(entity);
        // WHEN
        comp.save();
        tick(); // simulate async

        // THEN
        expect(service.update).toHaveBeenCalledWith(entity);
        expect(comp.isSaving).toEqual(false);
      }));

      it('Should call create service on save for new entity', fakeAsync(() => {
        // GIVEN
        const entity = new Concept();
        spyOn(service, 'create').and.returnValue(of(new HttpResponse({ body: entity })));
        comp.updateForm(entity);
        // WHEN
        comp.save();
        tick(); // simulate async

        // THEN
        expect(service.create).toHaveBeenCalledWith(entity);
        expect(comp.isSaving).toEqual(false);
      }));
    });
  });
});
